IO for clock reduction
