+++ benches/clock_reduction_bench.rs
@@ -1,7 +1,10 @@
 use criterion::{criterion_group, criterion_main, Criterion};
+use reveaal::ComponentLoader;
+
+mod bench_helper;
 use reveaal::extract_system_rep::create_executable_query;
 use reveaal::parse_queries::parse_to_query;
-use reveaal::{tests::TEST_SETTINGS, JsonProjectLoader, DEFAULT_SETTINGS};
+use reveaal::ProtobufServer::services::query_request::Settings;
 
 const QUERY: &str = "refinement: (((((Adm2 && HalfAdm1 && HalfAdm2) || Machine || Researcher) && ((Adm2 && HalfAdm1) || Machine || Researcher) && ((Adm2 && HalfAdm2) || Machine || Researcher) && ((HalfAdm1 && HalfAdm2) || Machine || Researcher) && (Adm2 || Machine || Researcher)) // (Adm2 && HalfAdm1 && HalfAdm2)) // Researcher) <= (((((Adm2 && HalfAdm1 && HalfAdm2) || Machine || Researcher) && ((Adm2 && HalfAdm1) || Machine || Researcher) && ((Adm2 && HalfAdm2) || Machine || Researcher) && ((HalfAdm1 && HalfAdm2) || Machine || Researcher) && (Adm2 || Machine || Researcher)) // (Adm2 && HalfAdm1 && HalfAdm2)) // Researcher)";
 
@@ -9,32 +12,31 @@ const QUERY: &str = "refinement: (((((Adm2 && HalfAdm1 && HalfAdm2) || Machine |
 /// The bench takes about 40 min on my machine, so grab some coffee.
 fn bench_clock_reduced_refinement(c: &mut Criterion) {
     // Set up the bench.
+    let loader = bench_helper::get_loader();
     let mut group = c.benchmark_group("Clock Reduction");
     group.bench_function("Refinement check - No reduction", |b| {
-        b.iter(normal_refinement);
+        loader.get_settings_mut().disable_clock_reduction = true;
+        b.iter(|| normal_refinement(loader));
     });
     group.bench_function("Refinement check - With reduction", |b| {
-        b.iter(clock_reduced_refinement);
+        loader.get_settings_mut().disable_clock_reduction = false;
+        b.iter(|| clock_reduced_refinement(loader));
     });
     group.finish();
 }
 
-fn clock_reduced_refinement() {
+fn clock_reduced_refinement(loader: &mut Box<dyn ComponentLoader>) {
     let query = parse_to_query(QUERY);
-    let mut loader =
-        JsonProjectLoader::new_loader("samples/json/EcdarUniversity".to_string(), DEFAULT_SETTINGS)
-            .to_comp_loader();
-    let executor = create_executable_query(query.get(0).unwrap(), &mut *loader).unwrap();
-    executor.execute();
+    create_executable_query(query.get(0).unwrap(), loader.as_mut())
+        .unwrap()
+        .execute();
 }
 
-fn normal_refinement() {
+fn normal_refinement(loader: &mut Box<dyn ComponentLoader>) {
     let query = parse_to_query(QUERY);
-    let mut loader =
-        JsonProjectLoader::new_loader("samples/json/EcdarUniversity".to_string(), TEST_SETTINGS)
-            .to_comp_loader();
-    let executor = create_executable_query(query.get(0).unwrap(), &mut *loader).unwrap();
-    executor.execute();
+    create_executable_query(query.get(0).unwrap(), loader.as_mut())
+        .unwrap()
+        .execute();
 }
 
 criterion_group! {
