+++ src/DataReader/json_writer.rs
@@ -1,4 +1,4 @@
-use crate::ModelObjects::Component;
+use crate::ModelObjects::component::Component;
 use std::fs::File;
 
 pub fn component_to_json_file(project_path: &str, component: &Component) {
