+++ src/DataReader/xml_parser.rs
@@ -1,9 +1,8 @@
 use crate::DataReader::parse_edge;
 use crate::DataReader::parse_edge::Update;
-use crate::ModelObjects::{
-    Component, Declarations, Edge, Location, LocationType, Query, SyncType, SystemDeclarations,
-    SystemSpecification,
-};
+use crate::ModelObjects::component::{Declarations, Edge, LocationType, SyncType};
+use crate::ModelObjects::system_declarations::{SystemDeclarations, SystemSpecification};
+use crate::ModelObjects::{component, queries, system_declarations, Expressions};
 use edbm::util::constraints::ClockIndex;
 use elementtree::{Element, FindChildren};
 use std::collections::HashMap;
@@ -18,7 +17,11 @@ pub fn is_xml_project(project_path: &str) -> bool {
 ///Used to parse systems described in xml
 pub(crate) fn parse_xml_from_file(
     fileName: &str,
-) -> (Vec<Component>, SystemDeclarations, Vec<Query>) {
+) -> (
+    Vec<component::Component>,
+    system_declarations::SystemDeclarations,
+    Vec<queries::Query>,
+) {
     //Open file and read xml
     let file = File::open(fileName).unwrap();
     let reader = BufReader::new(file);
@@ -26,17 +29,29 @@ pub(crate) fn parse_xml_from_file(
     parse_xml(reader)
 }
 
-pub(crate) fn parse_xml_from_str(xml: &str) -> (Vec<Component>, SystemDeclarations, Vec<Query>) {
+pub(crate) fn parse_xml_from_str(
+    xml: &str,
+) -> (
+    Vec<component::Component>,
+    system_declarations::SystemDeclarations,
+    Vec<queries::Query>,
+) {
     let reader = BufReader::new(xml.as_bytes());
 
     parse_xml(reader)
 }
 
-fn parse_xml<R: Read>(xml_data: R) -> (Vec<Component>, SystemDeclarations, Vec<Query>) {
+fn parse_xml<R: Read>(
+    xml_data: R,
+) -> (
+    Vec<component::Component>,
+    system_declarations::SystemDeclarations,
+    Vec<queries::Query>,
+) {
     let root = Element::from_reader(xml_data).unwrap();
 
     //storage of components
-    let mut xml_components: Vec<Component> = vec![];
+    let mut xml_components: Vec<component::Component> = vec![];
 
     for xml_comp in root.find_all("template") {
         let declarations = match xml_comp.find("declaration") {
@@ -44,7 +59,7 @@ fn parse_xml<R: Read>(xml_data: R) -> (Vec<Component>, SystemDeclarations, Vec<Q
             None => parse_declarations(""),
         };
         let edges = collect_edges(xml_comp.find_all("transition"));
-        let comp = Component {
+        let comp = component::Component {
             name: xml_comp.find("name").unwrap().text().parse().unwrap(),
             declarations,
             locations: collect_locations(
@@ -69,10 +84,10 @@ fn parse_xml<R: Read>(xml_data: R) -> (Vec<Component>, SystemDeclarations, Vec<Q
     (xml_components, system_declarations, vec![])
 }
 
-fn collect_locations(xml_locations: FindChildren, initial_id: &str) -> Vec<Location> {
-    let mut locations: Vec<Location> = vec![];
+fn collect_locations(xml_locations: FindChildren, initial_id: &str) -> Vec<component::Location> {
+    let mut locations: Vec<component::Location> = vec![];
     for loc in xml_locations {
-        let location = Location {
+        let location = component::Location {
             id: loc.get_attr("id").unwrap().parse().unwrap(),
             invariant: match loc.find("label") {
                 Some(x) => match parse_edge::parse_guard(x.text()) {
@@ -94,9 +109,9 @@ fn collect_locations(xml_locations: FindChildren, initial_id: &str) -> Vec<Locat
 }
 
 fn collect_edges(xml_edges: FindChildren) -> Vec<Edge> {
-    let mut edges: Vec<Edge> = vec![];
+    let mut edges: Vec<component::Edge> = vec![];
     for e in xml_edges {
-        let mut guard: Option<crate::ModelObjects::Expressions::BoolExpression> = None;
+        let mut guard: Option<Expressions::BoolExpression> = None;
         let mut updates: Option<Vec<Update>> = None;
         let mut sync: String = "".to_string();
         for label in e.find_all("label") {
@@ -117,7 +132,7 @@ fn collect_edges(xml_edges: FindChildren) -> Vec<Edge> {
                 _ => {}
             }
         }
-        let edge = Edge {
+        let edge = component::Edge {
             id: "NotImplemented".to_string(), // We do not support edge IDs for XML right now.
             source_location: e
                 .find("source")
