+++ src/EdgeEval/updater.rs
@@ -2,7 +2,7 @@ use std::collections::HashMap;
 use std::fmt;
 
 use crate::DataReader::parse_edge;
-use crate::ModelObjects::Declarations;
+use crate::ModelObjects::component::Declarations;
 use crate::ModelObjects::Expressions::ArithExpression;
 use colored::Colorize;
 use edbm::util::constraints::ClockIndex;
