+++ src/ModelObjects/mod.rs
@@ -1,14 +1,5 @@
 pub mod Expressions;
-mod component;
-mod edge;
-mod location;
-mod queries;
-mod state;
-mod statepair;
-mod system_declarations;
-mod transition;
-
-pub use self::{
-    component::*, edge::*, location::*, queries::*, state::*, statepair::*, system_declarations::*,
-    transition::*,
-};
+pub mod component;
+pub mod queries;
+pub mod statepair;
+pub mod system_declarations;
