+++ src/ModelObjects/statepair.rs
@@ -12,7 +12,7 @@ pub struct StatePair {
 }
 
 impl StatePair {
-    pub fn from_locations(
+    pub fn create(
         dimensions: usize,
         locations1: LocationTree,
         locations2: LocationTree,
