+++ src/Simulation/decision.rs
@@ -1,5 +1,7 @@
-use crate::ModelObjects::{State, Transition};
-use crate::TransitionSystems::TransitionSystemPtr;
+use crate::{
+    component::{State, Transition},
+    TransitionSystems::TransitionSystemPtr,
+};
 
 /// Represent a decision in a any composition of components: In the current `state` [`State`] we have decided to take this `action` [`String`].
 #[derive(Debug, Clone)]
