+++ src/System/refine.rs
@@ -2,7 +2,9 @@ use edbm::zones::OwnedFederation;
 use log::{debug, info, log_enabled, trace, Level};
 
 use crate::DataTypes::{PassedStateList, PassedStateListExt, WaitingStateList};
-use crate::ModelObjects::{StatePair, Transition};
+use crate::ModelObjects::component::Transition;
+
+use crate::ModelObjects::statepair::StatePair;
 use crate::System::query_failures::RefinementFailure;
 use crate::TransitionSystems::{LocationTree, TransitionSystemPtr};
 use std::collections::HashSet;
@@ -137,7 +139,7 @@ pub fn check_refinement(sys1: TransitionSystemPtr, sys2: TransitionSystemPtr) ->
     let initial_locations_1 = initial_locations_1.unwrap();
     let initial_locations_2 = initial_locations_2.unwrap();
 
-    let mut initial_pair = StatePair::from_locations(
+    let mut initial_pair = StatePair::create(
         dimensions,
         initial_locations_1.clone(),
         initial_locations_2.clone(),
@@ -161,10 +163,7 @@ pub fn check_refinement(sys1: TransitionSystemPtr, sys2: TransitionSystemPtr) ->
 
             let output_transition1 = sys1.next_outputs(curr_pair.get_locations1(), output);
             let output_transition2 = if extra {
-                vec![Transition::without_id(
-                    curr_pair.get_locations2(),
-                    dimensions,
-                )]
+                vec![Transition::new(curr_pair.get_locations2(), dimensions)]
             } else {
                 sys2.next_outputs(curr_pair.get_locations2(), output)
             };
@@ -183,10 +182,7 @@ pub fn check_refinement(sys1: TransitionSystemPtr, sys2: TransitionSystemPtr) ->
             let extra = extra_inputs.contains(input);
 
             let input_transitions1 = if extra {
-                vec![Transition::without_id(
-                    curr_pair.get_locations1(),
-                    dimensions,
-                )]
+                vec![Transition::new(curr_pair.get_locations1(), dimensions)]
             } else {
                 sys1.next_inputs(curr_pair.get_locations1(), input)
             };
