+++ src/TransitionSystems/common.rs
@@ -7,12 +7,11 @@ use edbm::{
 };
 use log::warn;
 
-use crate::ModelObjects::{Declarations, State, Transition};
-use crate::System::{
-    query_failures::{ConsistencyResult, DeterminismResult},
-    specifics::SpecificLocation,
+use crate::{
+    ModelObjects::component::{Declarations, State, Transition},
+    System::{query_failures::DeterminismResult, specifics::SpecificLocation},
 };
-use crate::TransitionSystems::CompositionType;
+use crate::{System::query_failures::ConsistencyResult, TransitionSystems::CompositionType};
 
 use super::{LocationTree, TransitionSystem, TransitionSystemPtr};
 
@@ -123,7 +122,7 @@ impl<T: ComposedTransitionSystem> TransitionSystem for T {
             return None;
         }
 
-        Some(State::new(init_loc, zone))
+        Some(State::create(init_loc, zone))
     }
 
     fn get_children(&self) -> (&TransitionSystemPtr, &TransitionSystemPtr) {
