+++ src/TransitionSystems/compiled_component.rs
@@ -1,4 +1,6 @@
-use crate::ModelObjects::{Component, DeclarationProvider, Declarations, State, Transition};
+use crate::ModelObjects::component::{
+    Component, DeclarationProvider, Declarations, State, Transition,
+};
 use crate::System::local_consistency::{self};
 use crate::System::query_failures::{
     ActionFailure, ConsistencyResult, DeterminismResult, SystemRecipeFailure,
@@ -62,7 +64,7 @@ impl CompiledComponent {
 
         for edge in component.get_edges() {
             let id = LocationID::Simple(edge.source_location.clone());
-            let transition = Transition::from_component_and_edge(&component, edge, dim);
+            let transition = Transition::from(&component, edge, dim);
             location_edges
                 .get_mut(&id)
                 .unwrap()
@@ -123,11 +125,11 @@ impl TransitionSystem for CompiledComponent {
         let is_input = self.inputs_contain(action);
 
         if locations.is_universal() {
-            return vec![Transition::without_id(locations, self.dim)];
+            return vec![Transition::new(locations, self.dim)];
         }
 
         if locations.is_inconsistent() && is_input {
-            return vec![Transition::without_id(locations, self.dim)];
+            return vec![Transition::new(locations, self.dim)];
         }
 
         let mut transitions = vec![];
