+++ src/TransitionSystems/location_tree.rs
@@ -1,7 +1,9 @@
 use edbm::{util::constraints::ClockIndex, zones::OwnedFederation};
 
-use crate::EdgeEval::constraint_applyer::apply_constraints_to_state;
-use crate::ModelObjects::{Declarations, Location, LocationType};
+use crate::{
+    EdgeEval::constraint_applyer::apply_constraints_to_state,
+    ModelObjects::component::{Declarations, Location, LocationType},
+};
 
 use super::LocationID;
