+++ src/lib.rs
@@ -14,8 +14,11 @@ pub use crate::DataReader::component_loader::{
     ComponentLoader, JsonProjectLoader, ProjectLoader, XmlProjectLoader,
 };
 pub use crate::DataReader::{parse_queries, xml_parser};
+pub use crate::ModelObjects::queries::Query;
 use crate::ProtobufServer::services::query_request::Settings;
 pub use crate::System::extract_system_rep;
+pub use ModelObjects::component;
+pub use ModelObjects::queries;
 pub use ProtobufServer::start_grpc_server_with_tokio;
 
 /// The default settings
