+++ src/tests/Simulation/helper.rs
@@ -9,23 +9,24 @@ use crate::ProtobufServer::services::{
     SimulationStepResponse,
 };
 
-impl SimulationStartRequest {
-    pub fn new(component_names: &[&str], components_path: &str, composition: &str) -> Self {
-        let simulation_info =
-            ProtoSimulationInfo::new(component_names, components_path, composition);
-        SimulationStartRequest {
-            simulation_info: Some(simulation_info),
-        }
+pub fn create_start_request(
+    component_names: &[&str],
+    components_path: &str,
+    composition: &str,
+) -> SimulationStartRequest {
+    let simulation_info = create_simulation_info_1(component_names, components_path, composition);
+    SimulationStartRequest {
+        simulation_info: Some(simulation_info),
     }
 }
 
-pub fn construct_step_requests(
+pub fn create_step_requests(
     component_names: &[&str],
     components_path: &str,
     composition: &str,
     last_response: Result<Response<SimulationStepResponse>, Status>,
 ) -> impl Iterator<Item = SimulationStepRequest> {
-    let simulation_info = ProtoSimulationInfo::new(component_names, components_path, composition);
+    let simulation_info = create_simulation_info_1(component_names, components_path, composition);
     let last_response = last_response.unwrap().into_inner();
     last_response
         .new_decision_points
@@ -36,28 +37,30 @@ pub fn construct_step_requests(
         })
 }
 
-impl ProtoSimulationInfo {
-    fn new(component_names: &[&str], components_path: &str, composition: &str) -> Self {
-        let json_components: Vec<_> = component_names
-            .iter()
-            .map(|component_name| ProtoComponent {
-                rep: Some(Rep::Json(
-                    fs::read_to_string(format!(
-                        "{}/Components/{}.json",
-                        components_path, component_name
-                    ))
-                    .unwrap(),
-                )),
-            })
-            .collect();
+fn create_simulation_info_1(
+    component_names: &[&str],
+    components_path: &str,
+    composition: &str,
+) -> ProtoSimulationInfo {
+    let json_components: Vec<_> = component_names
+        .iter()
+        .map(|component_name| ProtoComponent {
+            rep: Some(Rep::Json(
+                fs::read_to_string(format!(
+                    "{}/Components/{}.json",
+                    components_path, component_name
+                ))
+                .unwrap(),
+            )),
+        })
+        .collect();
 
-        ProtoSimulationInfo {
-            user_id: 0,
-            component_composition: composition.to_string(),
-            components_info: Some(ProtoComponentsInfo {
-                components: json_components,
-                components_hash: 0,
-            }),
-        }
+    ProtoSimulationInfo {
+        user_id: 0,
+        component_composition: composition.to_string(),
+        components_info: Some(ProtoComponentsInfo {
+            components: json_components,
+            components_hash: 0,
+        }),
     }
 }
