+++ src/tests/grpc/send_query.rs
@@ -15,7 +15,7 @@ mod refinements {
     #[tokio::test]
     async fn send_self_refinement_query() {
         let backend = ConcreteEcdarBackend::default();
-        let query_request = construct_query_request("refinement: Machine <= Machine");
+        let query_request = create_query_request("refinement: Machine <= Machine");
 
         let query_response = backend.send_query(query_request).await;
         assert!(query_response.is_ok());
@@ -31,7 +31,7 @@ mod refinements {
     #[tokio::test]
     async fn send_consistency_query() {
         let backend = ConcreteEcdarBackend::default();
-        let query_request = construct_query_request("consistency: Machine");
+        let query_request = create_query_request("consistency: Machine");
 
         let query_response = backend.send_query(query_request).await;
 
@@ -46,7 +46,7 @@ mod refinements {
     #[tokio::test]
     async fn send_determinism_query() {
         let backend = ConcreteEcdarBackend::default();
-        let query_request = construct_query_request("determinism: Machine");
+        let query_request = create_query_request("determinism: Machine");
 
         let query_response = backend.send_query(query_request).await;
 
@@ -58,7 +58,7 @@ mod refinements {
         }
     }
 
-    fn construct_query_request(query: &str) -> Request<QueryRequest> {
+    fn create_query_request(query: &str) -> Request<QueryRequest> {
         let json =
             std::fs::read_to_string(format!("{}/Components/Machine.json", ECDAR_UNI)).unwrap();
 
@@ -80,7 +80,7 @@ mod refinements {
     #[tokio::test]
     async fn send_panic_query() {
         let backend = ConcreteEcdarBackend::default();
-        let query_request = construct_query_request("refinement: Machine | <= Machine");
+        let query_request = create_query_request("refinement: Machine | <= Machine");
 
         let query_response = backend.send_query(query_request).await;
         assert!(query_response.is_err());
@@ -91,11 +91,11 @@ mod refinements {
     async fn send_after_panic_query() {
         let backend = ConcreteEcdarBackend::default();
         for _ in 0..5 {
-            let query_request = construct_query_request("refinement: Machine | <= Machine");
+            let query_request = create_query_request("refinement: Machine | <= Machine");
             let query_response = backend.send_query(query_request).await;
             assert!(query_response.is_err());
         }
-        let query_request = construct_query_request("refinement: Machine <= Machine");
+        let query_request = create_query_request("refinement: Machine <= Machine");
 
         let query_response = backend.send_query(query_request).await;
         assert!(query_response.is_ok());
