+++ src/tests/reachability/partial_state.rs
@@ -1,8 +1,8 @@
 #[cfg(test)]
 mod reachability_partial_states_test {
-    use crate::ModelObjects::{Declarations, Location, LocationType};
+    use crate::component::{Declarations, Location};
     use crate::TransitionSystems::CompositionType;
-    use crate::TransitionSystems::LocationTree;
+    use crate::{component::LocationType, TransitionSystems::LocationTree};
     use test_case::test_case;
 
     fn build_location_tree_helper(id: &str, location_type: LocationType) -> LocationTree {
